+++ ios/Classes/CountlyiOS/Countly.h
@@ -31,6 +31,10 @@ NS_ASSUME_NONNULL_BEGIN
  */
 - (void)startWithConfig:(CountlyConfig *)config;
 
+
+
+#pragma mark - Override Configuration
+
 /**
  * Sets a new host to be used in requests.
  * @discussion Requests already queued previously will also be using the new host.
@@ -41,12 +45,6 @@ NS_ASSUME_NONNULL_BEGIN
 - (void)setNewHost:(NSString *)newHost;
 
 /**
- * Sets a new URL session configuration to be used with all requests.
- * @param newURLSessionConfiguration The new URL session configuration
- */
-- (void)setNewURLSessionConfiguration:(NSURLSessionConfiguration *)newURLSessionConfiguration;
-
-/**
  * Sets a new app key to be used in new requests.
  * @discussion Before switching to the new app key, this method suspends Countly and resumes it immediately after.
  * @discussion The requests already queued prior to this method call will keep using the old app key.
@@ -55,6 +53,16 @@ NS_ASSUME_NONNULL_BEGIN
  * @param newAppKey The new app key
  */
 - (void)setNewAppKey:(NSString *)newAppKey;
+
+/**
+ * Sets a new URL session configuration to be used with all requests.
+ * @param newURLSessionConfiguration The new URL session configuration
+ */
+- (void)setNewURLSessionConfiguration:(NSURLSessionConfiguration *)newURLSessionConfiguration;
+
+
+
+#pragma mark - Queue Operations
 
 /**
  * Flushes request and event queues.
@@ -76,6 +84,19 @@ NS_ASSUME_NONNULL_BEGIN
  * @discussion these requests will be removed from request queue.
  */
 - (void)removeDifferentAppKeysFromQueue;
+
+/**
+ * Adds a direct request to the queue using given key-value pairs as query string pairs.
+ * @discussion requestParameters should be an @c NSDictionary, with keys and values are both @c NSString's only.
+ * @discussion Calls to this method will be ignored if:
+ * @discussion - There are not any consents given while @c requiresConsent flag is set on initial configuration.
+ * @param requestParameters Query string key-value pairs to be used in direct request
+ */
+- (void)addDirectRequest:(NSDictionary<NSString *, NSString *> * _Nullable)requestParameters;
+
+
+
+#pragma mark - Sessions
 
 /**
  * Starts session and sends @c begin_session request with default metrics for manual session handling.
@@ -392,17 +413,61 @@ NS_ASSUME_NONNULL_BEGIN
 #pragma mark - Crash Reporting
 
 /**
+ * Records an non-fatal exception.
+ * @discussion A convenience method for @c recordException:isFatal:stackTrace:segmentation:
+ * with isFatal passed as @c NO, stack trace and segmentation are passed as @c nil.
+ * @param exception Exception to be recorded
+ */
+- (void)recordException:(NSException *)exception;
+
+/**
+ * Records an exception with fatality information.
+ * @discussion A convenience method for @c recordException:isFatal:stackTrace:segmentation:
+ * with stack trace and segmentation are passed as @c nil.
+ * @param isFatal Whether the exception is fatal or not
+ */
+- (void)recordException:(NSException *)exception isFatal:(BOOL)isFatal;
+
+/**
+ * Records an exception with fatality information, given stack trace and segmentation.
+ * @discussion For manually recording all exceptions, fatal or not, with an ability to pass custom stack trace and segmentation data.
+ * @param exception Exception to be recorded
+ * @param isFatal Whether the exception is fatal or not
+ * @param stackTrace Stack trace to be recorded
+ * @param segmentation Crash segmentation to override @c crashSegmentation set on initial configuration
+ */
+- (void)recordException:(NSException *)exception isFatal:(BOOL)isFatal stackTrace:(NSArray * _Nullable)stackTrace segmentation:(NSDictionary<NSString *, NSString *> * _Nullable)segmentation;
+
+/**
+ * Records a Swift error with given stack trace.
+ * @discussion For manually recording Swift errors with an ability to pass custom stack trace.
+ * @param errorName A name describing the error to be recorded, a non-zero length valid string
+ * @param stackTrace Stack trace to be recorded
+ */
+- (void)recordError:(NSString *)errorName stackTrace:(NSArray * _Nullable)stackTrace;
+
+/**
+ * Records a Swift error with fatality information, given stack trace and segmentation.
+ * @discussion For manually recording Swift errors with an ability to pass custom stack trace and segmentation data.
+ * @param errorName A name describing the error to be recorded, a non-zero length valid string
+ * @param isFatal Whether the error is fatal or not
+ * @param stackTrace Stack trace to be recorded
+ * @param segmentation Crash segmentation to override @c crashSegmentation set on initial configuration
+ */
+- (void)recordError:(NSString *)errorName isFatal:(BOOL)isFatal stackTrace:(NSArray * _Nullable)stackTrace segmentation:(NSDictionary<NSString *, NSString *> * _Nullable)segmentation;
+
+/**
  * Records a handled exception manually.
  * @param exception Exception to be recorded
  */
-- (void)recordHandledException:(NSException *)exception;
+- (void)recordHandledException:(NSException *)exception DEPRECATED_MSG_ATTRIBUTE("Use 'recordException:' method instead!");
 
 /**
  * Records a handled exception and given stack trace manually.
  * @param exception Exception to be recorded
  * @param stackTrace Stack trace to be recorded
  */
-- (void)recordHandledException:(NSException *)exception withStackTrace:(NSArray * _Nullable)stackTrace;
+- (void)recordHandledException:(NSException *)exception withStackTrace:(NSArray * _Nullable)stackTrace DEPRECATED_MSG_ATTRIBUTE("Use 'recordException:isFatal:stackTrace:segmentation:' method instead! (passing isFatal:NO, segmentation:nil)");
 
 /**
  * Records an unhandled exception and given stack trace manually.
@@ -410,7 +475,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param exception Exception to be recorded
  * @param stackTrace Stack trace to be recorded
  */
-- (void)recordUnhandledException:(NSException *)exception withStackTrace:(NSArray * _Nullable)stackTrace;
+- (void)recordUnhandledException:(NSException *)exception withStackTrace:(NSArray * _Nullable)stackTrace DEPRECATED_MSG_ATTRIBUTE("Use 'recordException:isFatal:stackTrace:segmentation:' method instead! (passing isFatal:YES, segmentation:nil)");
 
 /**
  * Records custom logs to be delivered with crash report.
@@ -484,20 +549,6 @@ NS_ASSUME_NONNULL_BEGIN
  */
 + (CountlyUserDetails *)user;
 
-/**
- * Handles switching from device ID to custom user ID for logged in users
- * @discussion When a user logs in, this user can be tracked with custom user ID instead of device ID.
- * @discussion This is just a convenience method that handles setting user ID as new device ID and merging existing data on Countly Server.
- * @param userID Custom user ID uniquely defining the logged in user
- */
-- (void)userLoggedIn:(NSString *)userID DEPRECATED_MSG_ATTRIBUTE("Use 'setNewDeviceID:onServer:' method instead!");
-
-/**
- * Handles switching from custom user ID to device ID for logged out users
- * @discussion When a user logs out, all the data can be tracked with default device ID henceforth.
- * @discussion This is just a convenience method that handles resetting device ID to default one and starting a new session.
- */
-- (void)userLoggedOut DEPRECATED_MSG_ATTRIBUTE("Use 'setNewDeviceID:onServer:' method instead!");
 
 
 
@@ -526,7 +577,7 @@ NS_ASSUME_NONNULL_BEGIN
  * @param widgetID ID of the feedback widget created on Countly Server.
  * @param completionHandler A completion handler block to be executed when feedback widget is dismissed by user or there is an error.
  */
-- (void)presentFeedbackWidgetWithID:(NSString *)widgetID completionHandler:(void (^)(NSError * __nullable error))completionHandler DEPRECATED_MSG_ATTRIBUTE("Use 'presentRatingWidgetWithID:' method instead!");
+- (void)presentFeedbackWidgetWithID:(NSString *)widgetID completionHandler:(void (^)(NSError * __nullable error))completionHandler DEPRECATED_MSG_ATTRIBUTE("Use 'presentRatingWidgetWithID:completionHandler' method instead!");
 
 /**
  * Presents rating widget with given ID in a WKWebView placed in a UIViewController.
@@ -599,7 +650,7 @@ NS_ASSUME_NONNULL_BEGIN
 
 /**
  * Records indirect attribution with given key-value pairs.
- * @discussion Keys could be a predefined CLYAttributionKey or any non-zero length valid string.
+ * @discussion Keys could be a predefined @c CLYAttributionKey or any non-zero length valid string.
  * @discussion This method sends an immediate request.
  * @discussion Calls to this method will be ignored if:
  * @discussion - Consent for @c CLYConsentAttribution is not given, while @c requiresConsent flag is set on initial configuration.